Add cross boundary result helpers
