+++ lib/ain-rs-exports/src/core.rs
@@ -1,13 +1,14 @@
 use crate::ffi::CrossBoundaryResult;
+use crate::prelude::*;
 
 pub fn ain_rs_preinit(result: &mut CrossBoundaryResult) {
     ain_grpc::preinit();
-    result.ok = true;
+    cross_boundary_success(result);
 }
 
 pub fn ain_rs_init_logging(result: &mut CrossBoundaryResult) {
     ain_grpc::init_logging();
-    result.ok = true;
+    cross_boundary_success(result);
 }
 
 pub fn ain_rs_init_core_services(result: &mut CrossBoundaryResult) {
@@ -17,7 +18,7 @@ pub fn ain_rs_init_core_services(result: &mut CrossBoundaryResult) {
 
 pub fn ain_rs_stop_core_services(result: &mut CrossBoundaryResult) {
     ain_grpc::stop_services();
-    result.ok = true;
+    cross_boundary_success(result);
 }
 
 pub fn ain_rs_init_network_services(
@@ -26,24 +27,14 @@ pub fn ain_rs_init_network_services(
     grpc_addr: &str,
 ) {
     match ain_grpc::init_network_services(json_addr, grpc_addr) {
-        Ok(()) => {
-            result.ok = true;
-        }
-        Err(e) => {
-            result.ok = false;
-            result.reason = e.to_string();
-        }
+        Ok(_) => cross_boundary_success(result),
+        Err(e) => cross_boundary_error_return(result, e.to_string()),
     }
 }
 
 pub fn ain_rs_stop_network_services(result: &mut CrossBoundaryResult) {
     match ain_grpc::stop_network_services() {
-        Ok(()) => {
-            result.ok = true;
-        }
-        Err(e) => {
-            result.ok = false;
-            result.reason = e.to_string();
-        }
+        Ok(_) => cross_boundary_success(result),
+        Err(e) => cross_boundary_error_return(result, e.to_string()),
     }
 }
