+++ lib/ain-rs-exports/src/lib.rs
@@ -1,5 +1,6 @@
 mod core;
 mod evm;
+mod prelude;
 
 use crate::core::*;
 use crate::evm::*;
